fixup! fix(sql): treat zero char as empty string literal
